build: bump svelkit vite
